Add decimal functionality

+++ Python Excercises/rectangleAreaCalc.py
@@ -1,5 +1,5 @@
-length = int(input("Enter the length: "))
-width = int(input("Enter the width: "))
+length = float(input("Enter the length: "))
+width = float(input("Enter the width: "))
 
 area = length * width
